+++ mex/editor/ingest/models.py
@@ -3,19 +3,12 @@ import reflex as rx
 from mex.editor.models import EditorValue
 
 
-class AuxResult(rx.Base):
-    """Auxiliary search result."""
+class IngestResult(rx.Base):
+    """Ingest search result."""
 
     identifier: str
     title: list[EditorValue]
     preview: list[EditorValue]
     show_properties: bool
     all_properties: list[EditorValue]
-    show_import_button: bool
-
-
-class AuxNavItem(rx.Base):
-    """Model for one aux navigation bar item."""
-
-    title: str = ""
-    value: str = "/"
+    show_ingest_button: bool