+++ CHANGELOG.md
@@ -8,8 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Test for MEx-Editor Rules transform
+- fields.py as look-up for required fields for all models
 
 ### Changes
+- required field names for edit mode from 'REQUIRED_FIELDS_BY_CLASS_NAME' in fields.py
+- test_trasnform.py refactored because EditorField has now attribute 'is_required'
+- consent preview uses direct field isIndicatedAtTime as title
 
 ### Deprecated
 
@@ -18,6 +23,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [0.18.1] - 2025-06-19
+
+### Changes
+
+- rename aux-search component to ingest, matching naming in backend
+
+### Fixed
+
+- fix transparency of overlay containers
+
+## [0.18.0] - 2025-06-17
+
+### Changes
+
+- install mex-common and mex-artificial from pypi
 
 ## [0.17.0] - 2025-06-16